6309366 gcc and cmd/cvcd don't get along 6272090 gcc and cmd/pcmciad don't get along 6273814 gcc and cmd/rmformat don't get along 6273880 gcc and cmd/smserverd don't get along 6309225 gcc and efcode don't get along 6309403 gcc and se don't get along 6310538 gcc and stp4020 don't get along 6309970 gcc and fas don't get along 6310423 gcc and zs don't get along 6309951 gcc and ebus don't get along 6310427 gcc and su driver don't get along 6310552 gcc and starfire don't get along 6308013 gcc and cmd/picl don't get along 6310441 gcc and rmclomv don't get along 6309943 gcc and sun ata don't get along 6309421 gcc and excalibur/io don't get along

+++ usr/src/lib/efcode/engine/debug.c
@@ -20,7 +20,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2000-2003 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -812,7 +812,7 @@ dump_line(uchar_t *ptr)
 	uchar_t *byte;
 	int i;
 
-	log_message(MSG_INFO, "%p  ", (uint32_t)ptr);
+	log_message(MSG_INFO, "%p  ", ptr);
 	for (i = 0, byte = ptr; i < 16; i++) {
 		if (i == 8)
 			log_message(MSG_INFO, " ");
@@ -1048,7 +1048,7 @@ dump(fcode_env_t *env)
 	offset = ((long)data) & 0xf;
 	len += offset;
 	data = (uchar_t *)((long)data & ~0xf);
-	sprintf(buf, "%p", (uint32_t)data);
+	sprintf(buf, "%p", data);
 	print_bytes_header(strlen(buf), offset);
 	for (len += offset; len > 0; len -= 16, data += 16)
 		dump_line(data);
